start of work to add additional graph outputs and custom packages
we cannot always rely on libraries.io to have all the package managers we need! E.g., spack or github
can be used as sources, and we should be able to interact with their respective APIs. This refactor
takes that into account. It increases the complexity of the design to be able to support both
these ways but we should be okay with enough docs/comments/tests.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>