start of work to refactor strucutre (#16)
* start of work to refactor structure
* ensure setting file is exported in environment
* move http and requests into own package
* adding example to parse (and combine) repos

I want the settings and cache to be more global, e.g., easily accessible from anywhere
and not reliant on a central client! This is start of work to enable that. I need to
add cache support next to the custom package parser, and then see if I can remove
the endpoint functions from the client and make general "use from anywhere" functions

Signed-off-by: vsoch <vsoch@users.noreply.github.com>