adding support to use pypi resolver to get dependency list from file (#29)
* adding support to use pypi resolver to get dependency list from file
* allow loading from cached data

as suggested by @alecbcs, this approach can be slightly more accurate! It works very well for requirements.txt files that can
exist outside of the context ofa repo, however for setup.py it is ideally run in context of the repository,
as that context is needed for success. Either way, if an attempt is not successful we fall back to the static parser.
I am running it now across the entire rsepedia Python set and will do a final review / merge when everything looks good

Signed-off-by: vsoch <vsoch@users.noreply.github.com>